42_cursus/Transcendence/projet_transcendence/the_transcendence/requirements.txt
asgiref==3.8.1
Django==5.1.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-allauth==65.3.1
PyJWT==2.10.1
pillow==11.1.0
psycopg2-binary==2.9.10
requests==2.32.3
sqlparse==0.5.2
#gunicorn==23.0.0
#djangorestframework_simplejwt==5.4.0